gtp_bitslide: prevent bitslide from going out of range
If, for some reason, a proper alignment is missed then the cur_slide counter
might increment outside its window (0..9 for an 8 bit Phy and 0..19 for a 16 bit
Phy). Apparently nobody ever missed an alignment, but it might fail someday,
somehow, sometime... Therefore it is wise to put the module function in place.
The module function was implemented using an if statement because synthesis
tools complained about the VHDL "mod" function.