altera_phase: fix signal transposition bug
By chance we were always using altera_phase with 3 outputs.
The aligned and sync_rstn matrices were thus square 3x3.

When using more outputs, it became clear the types were wrong and
had set the columns to be rows. No logic was changed; only the shape
of the matrix of signals.