wr_endpoint/rx_path: don't store pfilter decision if frame is dropped by rx_buffer
If there is a storm of frames and rx_buffer gets full it starts to drop frames.
However, before this commit even if a frame was dropped that way, U_match_buffer
was storing pfilter decision for that frame. As a result decisions coming out
from the match_buffer were shifted and applied to wrong frames. With this commit
I check if the rx_buffer is full at the beginning of frame and eventually block
mbuf_we so that the pfilter decision is not stored.