wr_core: updating the core and a reference design to feed phy_rdy_i of the endpoint
phy_rdy_i was needed to fix the Virtex-6 FIFO reseting issue on the WRS.
However, WRPC uses inferred fifo which has fully asynchronous reset and does not
suffer from this bug. Therefore in case of the WRPC this signal can be tied to
'1' or drived from the GTP serdes.