modules: regenerate wishbone Slave interfaces with updated wbgen
Sometimes *.wb file had to be updated to produce the same result as used till
now. Earlier wbgen2 had a problem accepting _align_ for the first field in a
register. The bug was fixed, but we want to keep the alignment the old way
(already used in the software). Hence I had to delete some of the _align_
directives.