channel: increase valid counter on 1st cycle of strobe
Previously, the valid counter would increase on the last cycle of a strobe.
Thus, if it were stalled forever, the counter would not increase.  That
posed a problem for software driven queues which need the counter to
increase in order to receive an MSI to pop the queue via lowering stall.