[wr streamers] implemented snapshot of the statistis
it was hard to get fully coherent statistics, especially if they were changing
fast (e.g. 250kHz in btrain). Setting snapshot bit to high copies at the same
instant all the counters to separate registers. as long as snapshot bit is high,
these registers are exposed to WB and SNMP. so one can read it as long as it
takes and the data is still coherent.