modules/wr_pps_gen: add PPS mask/unmask bit in ESCR register.
By default, the bit is zero (masked), which means that PPS signal output is masked
when the link status is not ok. This is useful to pervent spurious PPS when in GM
or SLAVE mode and the link goes down.

When the bit is set, then a PPS is always generated (as long as the PPS valid bit
in the ESCR register is set). This is useful in free-running master mode.