[wr_streamers] fixed latency bugfix
When frames were sent to close to each other (high frequency), the
timestamp of incoming frame would reset the delay counter of the
previously received frame. This could potentially cause infinite
fixed delay... Now, in such case, the counter is not reseted and
so the following (too closely) frame will be exposed to the user
immediatelly.