[wr_streamers] added generic to set simulation startup timer value
when streamers are used in a simulation of top entity, the startup
timer is needed, thought it should be appropriate for the simulation
time. when streamers are simulated alone, the startup timer is
not needed. the added generic allows to set the timer (i.e. override
the default value to zero)